requirements.txt
discord.py
discord-ext-slash
aiofiles
aiosqlite
google-cloud-translate